# docs/swagger.yaml
definitions:
  api.LoginInfo:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  api.LoginResponse:
    properties:
      token:
        type: string
    type: object
  api.RegisterInfo:
    properties:
      email:
        type: string
      name:
        type: string
      password:
        type: string
      phone:
        type: string
    type: object
  api.RegisterResponse:
    properties:
      new_password:
        type: string
      old_password:
        type: string
    type: object
  api.UserInfoResponse:
    properties:
      avatar_path:
        type: string
      create_at:
        type: string
      email:
        type: string
      phone:
        type: string
      user_id:
        type: integer
      user_name:
        type: string
    type: object
info:
  contact: {}
  title: Kayak Backend API
  version: 0.0.1
paths:
  /login:
    post:
      description: 用户登录
      parameters:
      - description: 用户登陆信息
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/api.LoginInfo'
      responses:
        "200":
          description: 用户登陆反馈
          schema:
            $ref: '#/definitions/api.LoginResponse'
        "400":
          description: 请求解析失败
          schema:
            type: string
        default:
          description: 服务器错误
          schema:
            type: string
  /logout:
    get:
      description: 用户退出
      responses:
        "200":
          description: 退出成功
          schema:
            type: string
        default:
          description: 服务器错误
          schema:
            type: string
      security:
      - ApiKeyAuth: []
  /ping:
    get:
      description: 测试服务器是否正常运行
      responses:
        "200":
          description: pong
          schema:
            type: string
        default:
          description: 服务器错误
          schema:
            type: string
  /register:
    post:
      description: 用户注册
      parameters:
      - description: 用户注册信息
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/api.RegisterInfo'
      responses:
        "200":
          description: 注册成功
          schema:
            type: string
        "400":
          description: 请求解析失败
          schema:
            type: string
        "409":
          description: 用户名已存在
          schema:
            type: string
        default:
          description: 服务器错误
          schema:
            type: string
  /reset-password:
    post:
      description: 用户修改密码
      parameters:
      - description: 用户修改密码信息
        in: body
        name: info
        required: true
        schema:
          $ref: '#/definitions/api.RegisterResponse'
      responses:
        "200":
          description: 修改成功
          schema:
            type: string
        "400":
          description: 请求解析失败
          schema:
            type: string
        "409":
          description: 用户不存在
          schema:
            type: string
        default:
          description: 服务器错误
          schema:
            type: string
      security:
      - ApiKeyAuth: []
  /user/info:
    get:
      description: 获取用户信息
      responses:
        "200":
          description: 用户信息
          schema:
            $ref: '#/definitions/api.UserInfoResponse'
        default:
          description: 服务器错误
          schema:
            type: string
      security:
      - ApiKeyAuth: []
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: X-Token
    type: apiKey
swagger: "2.0"